Do not use pathlib for hdf5 datasets, fixes #2318
