+++ docs/changes/2319.bugfix.rst
@@ -1,0 +1,1 @@
+Fix HDF5Writer not working on windows due to using pathlib for hdf5 dataset names.
